Ported for 24-bit adc/dac

+++ signalreader/InputSignalReader.h
@@ -37,14 +37,14 @@ private:
     snd_pcm_uframes_t bufferSize;
     snd_pcm_t *pcm_handle;
 
-    function<void(snd_pcm_uframes_t, int16_t*)> callback;
+    function<void(snd_pcm_uframes_t, int32_t*)> callback;
 
     void readSamples();
 
 public:
     InputSignalReader();
 
-    void registerCallback(function<void(snd_pcm_uframes_t, int16_t*)> callback);
+    void registerCallback(function<void(snd_pcm_uframes_t, int32_t*)> callback);
 
     void start();
     void stop();
